Re-define coverage for collateData()
Participation is defined by sum of junction counts at event bounds divided by the total SpliceOver of these bounds.
Estimate is now derived as the larger of the two alternate events (rather than defining exon cluster groups).
The reason for doing this is because there are many instances where junctions are bound by both flanks in the same exon cluster (as a result of retained intron annotations). Calculation of these are not consistent.

Coverage for retained introns are derived from that of intron coverages from processBAM output (as it should!)