Improved novel ASE detection to reduce false positives
A simple filtering strategy is implemented in collateData() to reduce the number of novel junctions prior to assembly of the novel ASE reference.

Novel junctions are only used for novel ASE reference if it is detected (>= 1 reads) in at least 3 samples, or if there are at least 10 reads in at least 1 sample. These settings can be customized - see ?collateData